Use runners from vars if available (#812)
This was removed in #801 to bring the file in line with downstream.
However, it is better to rely on the vars environment when invoking the
main workflow from dispatch, while offering an override when invoking
the workflow as a sub-workflow.

(cherry picked from commit 67bb06c4658157f06e034333426f83504bfe5eb1)